backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.31
pydantic==2.7.4
pydantic-settings==2.3.3
psycopg2-binary==2.9.9
cryptography==42.0.8
httpx==0.27.0
openai==1.40.0
google-api-python-client==2.140.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
